building main application
