--- mern-structure/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "mern-structure",
-  "version": "1.0.0",
-  "description": "MERNstack app",
-  "main": "server.js",
-  "scripts": {
-    "start": "node server.js",
-        "server": "nodemon server.js",
-        "client": "cd client && npm start ",
-        "clientinstall": " cd client && npm install ",
-        "dev": "concurrently -n 'server,client' -c 'red,green' \"npm run server\" \"npm run client\"",
-        "heroku-postbuild": "NPM_CONFIG_PRODUCTION=false npm install --prefix client && npm run build --prefix client"
-  },
-  "author": "Agrit Tiwari",
-  "license": "MIT",
-  "devDependencies": {
-      "concurrently": "^5.3.0",
-      "nodemon": "^2.0.4"
-  }
-}
