+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "nextjs-structure",
+  "name": "nextjs-auth0-tailwind-todo",
   "private": true,
   "scripts": {
     "dev": "next dev",
