+++ pages/index.js
@@ -1,0 +1,31 @@
+import Head from 'next/head'
+import Image from 'next/image'
+import styles from '../styles/Home.module.css'
+
+export default function Home() {
+  return (
+    <div className={styles.container}>
+      <Head>
+        <title>Create Next App</title>
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <div>
+        <div>Client Side Rendering</div>
+        <div><ul>Server side Rendering</ul></div>
+        <h2>Static Side Generation</h2>
+</div>
+
+      <footer className={styles.footer}>
+        <a
+          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          A Devkode member 🔥 
+          
+        </a>
+      </footer>
+    </div>
+  )
+}
